Move cases to separate file.
